# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateTodoInput:
    properties:
      task:
        example: Swagger 문서 수정하기
        type: string
    required:
    - task
    type: object
  model.Todo:
    properties:
      created_at:
        description: 생성 시간
        type: string
      done:
        type: boolean
      id:
        description: |-
          gorm.Model 대신 필요한 것만 직접 정의합니다.
          `gorm:"primaryKey"` 태그로 이게 PK임을 알려줍니다.
        type: integer
      task:
        type: string
    type: object
  model.WebResponse:
    properties:
      code:
        description: 결과 코드 (HTTP Status와 동일하게 쓰거나 커스텀 코드 사용)
        example: 200
        type: integer
      data:
        description: 실제 데이터 (Payload) - 무엇이든 들어갈 수 있음(Generic)
      message:
        description: 사람이 읽을 수 있는 메시지
        example: Success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: gyong97@example.com
    name: Gyong97
  description: 이것은 Go로 만든 Todo 리스트 API 문서입니다.
  title: Go Todo API
  version: "1.0"
paths:
  /dashboard:
    get:
      consumes:
      - application/json
      description: 사용자 정보와 할 일 통계를 병렬로 조회하여 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 대시보드 데이터 조회
      tags:
      - Dashboard
  /reports:
    post:
      consumes:
      - application/json
      description: 리포트 생성을 비동기로 요청합니다. (처리 결과는 이메일 발송 등)
      produces:
      - application/json
      responses:
        "202":
          description: 요청 접수됨
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: 일일 리포트 생성 요청
      tags:
      - Reports
  /todos:
    get:
      consumes:
      - application/json
      description: 저장된 모든 할 일 목록을 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Todo'
                  type: array
              type: object
      summary: 할 일 목록 조회
      tags:
      - Todos
    post:
      consumes:
      - application/json
      description: 새로운 할 일을 목록에 추가합니다.
      parameters:
      - description: 할 일 정보
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTodoInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Todo'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.WebResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: 할 일 추가
      tags:
      - Todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: 특정 ID의 할 일을 영구적으로 삭제합니다.
      parameters:
      - description: 삭제할 할 일 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 삭제 성공
          schema:
            $ref: '#/definitions/model.WebResponse'
        "400":
          description: 잘못된 ID 형식
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: ID를 찾을 수 없음
          schema:
            $ref: '#/definitions/model.WebResponse'
        "500":
          description: 서버 내부 에러
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: 할 일 삭제
      tags:
      - Todos
    patch:
      consumes:
      - application/json
      description: 특정 ID의 할 일 완료 상태(Done)를 반전시킵니다.
      parameters:
      - description: 수정할 할 일 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
        "400":
          description: 잘못된 ID 형식
          schema:
            $ref: '#/definitions/model.WebResponse'
        "404":
          description: ID를 찾을 수 없음
          schema:
            $ref: '#/definitions/model.WebResponse'
      summary: 할 일 완료 여부 토글
      tags:
      - Todos
swagger: "2.0"